Map resource load button wait
